UPDATED AND ADDED HOME PAGE, SEARCH FUNCTION AND FIXED CODE
