+++ client/src/components/Home/Home.css
@@ -1,77 +1,372 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;700&display=swap');
-
 * {
-  font-family: 'Poppins', sans-serif;
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  overflow-x: hidden;
 }
 
-.home-desc-header {
-  color: white;
+.iou-logo-txt {
+  color: #ffffff;
+  font-size: 40px;
+}
+
+.main-content {
+  width: 100vw;
+  height: 100vh;
+  position: relative;
+  overflow-x: hidden;
+}
+.content {
+  width: 100%;
+  height: 100%;
+}
+
+.content-txt {
+  color: #ffffff;
   font-size: 50px;
-  font-weight: 700;
-  margin-bottom: 0px;
-}
-
-.home-desc-body {
-  color: white;
-  display: inline-block;
-  font-size: 20px;
-  font-weight: 400;
-  width: 516px;
-}
-
-.home-img {
   position: absolute;
-  top: 0px;
-  left: 0px;
-  z-index: -1;
-  width: 872px;
-  height: 1073px;
+  top: 80px;
+  z-index: 50;
+  margin-top: 100px;
+  left: 50px;
+  text-align: left;
 }
 
-.home-grid {
-  display: grid;
-  grid-template-columns: 872px auto;
+.register-btn {
+  color: #8f2a7e;
+  font-size: 25px;
+  font-weight: bold;
+  background-color: #ffffff;
+  height: 60px;
+  position: absolute;
+  top: 100px;
+  z-index: 50;
+  margin-top: 330px;
+  left: 50px;
+  border-radius: 10px;
+  border: none;
+  width: 300px;
+  cursor: pointer;
 }
 
-.home-img-text {
-  margin-top: 80px;
+.public-favour-text-below {
+  margin-top: 20px;
+  color: #9b9b9b;
 }
 
-.home-register-title {
+.register-btn:hover {
+  transition-duration: 0.3s;
+  background-color: #ff6122;
+  color: #ffffff;
+}
+.register-btn:focus {
+  outline: none;
+  color: #ffffff;
+  background-color: #ff6122;
+}
+
+.content-txt-below {
+  color: #ffffff;
+  position: absolute;
+  top: 50px;
+  z-index: 50;
+  font-size: 20px;
+  margin-top: 300px;
+  left: 50px;
+  text-align: left;
+}
+
+.public-title {
   font-size: 50px;
 }
 
-.home-profile-icon {
-  width: 90px;
-  height: 90px;
+.navbar-top {
+  position: absolute;
+  left: 0;
+  top: 0;
+  width: 100vw;
+  height: 7.5vh;
+  /* box-shadow: 0 1px 1px rgba(0, 0, 0, 0.08), 0 2px 2px rgba(0, 0, 0, 0.08),
+    0 4px 4px rgba(0, 0, 0, 0.08), 0 8px 8px rgba(0, 0, 0, 0.08),
+    0 16px 16px rgba(0, 0, 0, 0.08); */
+  z-index: 50;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 30px;
+  font-family: 'Montserrat', sans-serif;
 }
 
-.sign-up-button {
-  display: inline-block;
+.nav-middle {
+  margin: 0% auto;
+}
+.nav-right {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  font-weight: 600;
+  width: 25%;
+  margin-right: 0%;
+}
+#home-txt {
+  margin-top: 1%;
+  border-bottom: 3px solid #8a2980;
+  cursor: pointer;
+}
+#home-txt:hover {
+  border-bottom: 3px solid #8a2980 !important;
+}
+
+#leaderboard-txt {
+  margin-top: 1%;
+  border-bottom: 3px solid #ffffff;
+  cursor: pointer;
+}
+
+#leaderboard-txt:hover {
+  border-bottom: 3px solid #8a29809a !important;
+}
+
+.signout-btn {
+  transition-duration: 0.3s;
+  font-family: 'Montserrat', sans-serif;
+  background: white;
+  border-radius: 5px;
+  width: 180px;
+  height: 30px;
+  border: 2px solid #0cbe88;
+  font-weight: 600;
+  color: #0cbe88;
+  cursor: pointer;
+}
+.signout-btn:hover {
+  transition-duration: 0.3s;
+  background-color: #0cbe88;
+  color: #ffffff;
+}
+.signout-btn:focus {
+  outline: none;
+  color: #ffffff;
+  background-color: #0cbe88;
+}
+.nav-logo {
+  height: 50px;
+  width: 50px;
+  margin-left: 50px;
+  margin-right: 20px;
+}
+.top-right {
+  vertical-align: middle;
+}
+
+#home-stack {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+  -webkit-animation: slide-down 0.4s ease-out;
+  -moz-animation: slide-down 0.4s ease-out;
+  animation: slide-down 0.4s ease-out;
+}
+#account-stack {
+  display: none;
+  -webkit-animation: slide-down 0.4s ease-out;
+  -moz-animation: slide-down 0.4s ease-out;
+  animation: slide-down 0.4s ease-out;
+}
+
+@-webkit-keyframes slide-down {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+@-moz-keyframes slide-down {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+@keyframes slide-down {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+
+.search-input {
+  background: #f2f2f2;
+  width: 550px;
+  height: 35px;
+  border: 2px solid #f2f2f2;
+  border-top-left-radius: 30px;
+  border-bottom-left-radius: 30px;
+  border-top-right-radius: 30px;
+  border-bottom-right-radius: 30px;
+  padding: 10px;
+  padding-left: 15px;
+  font-size: 14px;
+  box-shadow: none !important;
+  transition-duration: 0.7s;
+  caret-color: #8a2980;
+}
+.search-input:hover,
+.search-input:focus {
+  transition-duration: 0.7s;
+  outline: none;
+  border: 2px solid #8a2980;
+}
+.search-input:hover + .search-btn {
+  transition-duration: 0.7s;
+  outline: none;
+  border-top-left-radius: 0px;
+  border-bottom-left-radius: 0px;
+  background-color: #8a2980;
+}
+.search-input:focus + .search-btn {
+  transition-duration: 0.7s;
+  outline: none;
+  border-top-left-radius: 0px;
+  border-bottom-left-radius: 0px;
+  background-color: #8a2980;
+}
+.search-input:hover ~ .search-btn .search-icon {
+  color: #ffffff;
+}
+.search-input:focus ~ .search-btn .search-icon {
+  color: #ffffff;
+}
+
+.search-btn {
+  transition-duration: 0.7s;
+  height: 35px;
+  background-color: #ffffff;
+  border: 2px solid #8a2980;
+  padding: 0 15px;
+  border-top-left-radius: 30px;
+  border-bottom-left-radius: 30px;
+  border-top-right-radius: 30px;
+  border-bottom-right-radius: 30px;
+  font-size: 14px;
+  float: right;
+  position: absolute;
+  right: 0px;
+  width: 70px;
+  outline: 0;
+  cursor: pointer;
+}
+.search-btn:hover {
+  transition-duration: 0.15s;
+  background-color: #8a2980;
+}
+.search-btn:hover .search-icon {
+  color: #ffffff;
+}
+.search-form {
+  display: flex;
+  align-items: center;
+  position: relative;
+  justify-content: space-around;
+}
+
+.background-img {
+  position: relative;
+  z-index: 10;
+  left: 0;
+  top: 0;
+  overflow: hidden;
+}
+
+.isometric-img {
+  position: absolute;
+  z-index: 15;
+  right: 50px;
+  top: 100px;
+}
+
+.favours-card {
+  width: 1000px;
+  height: 100px;
+  margin: 50px auto;
+  border-radius: 10px;
+  border: 2px solid #dbdbdb;
+  -webkit-box-shadow: 0px 9px 20px 3px rgba(26, 26, 26, 0.16);
+  -moz-box-shadow: 0px 9px 20px 3px rgba(26, 26, 26, 0.16);
+  box-shadow: 0px 9px 20px 3px rgba(26, 26, 26, 0.16);
+}
+
+.favours-card-square-bg {
+  background-color: #ff6464;
+  height: 100%;
+  float: left;
+  width: 100px;
+  overflow: hidden;
+}
+
+.avatar {
+  width: 70px !important;
+  height: 100px !important;
+}
+
+.title-txt {
+  margin-top: 10px;
+  margin-left: 20px;
+  text-align: left;
+  float: left;
+  margin-right: 30px;
+  font-size: 15px;
+  color: #8b9ca8;
+}
+
+.card-text {
+  font-size: 17px;
+  color: #596773;
   font-weight: bold;
-  font-size: 25px;
+  display: inline;
+  float: left;
+  margin-left: 20px;
+}
+
+.points-text {
+  color: #15b987;
+  float: left;
+  text-align: left;
+  margin-left: 20px;
+  font-weight: bolder;
+}
+
+.do-favour-btn {
+  float: right;
+  display: inline;
+  transition-duration: 0.3s;
+  font-family: 'Montserrat', sans-serif;
+  background: #fe8e51;
+  border-radius: 5px;
+  width: 150px;
+  border: none;
+  margin-right: 40px;
+  height: 40px;
+  font-weight: 600;
   color: white;
-  padding: 10px 0px;
-  width: 452px;
-  background-color: #01c48a;
-  border-radius: 15px;
   cursor: pointer;
 }
 
-.home-account-text {
-  margin-top: 0px;
-  font-size: 20px;
-  cursor: pointer;
-}
-
-.cancel-button {
-  display: inline-block;
+.tag {
+  color: #4d6bff;
+  display: block;
+  border-radius: 30px;
   font-weight: bold;
-  font-size: 25px;
-  color: white;
-  padding: 10px 0px;
-  width: 452px;
-  background-color: #ec3030;
-  border-radius: 15px;
-  cursor: pointer;
+  margin-top: 8px;
+  width: 100px;
+  height: 25px;
+  background-color: #afd2ff;
+}
+
+.tag-content {
 }
