+++ client/src/components/SearchPage/SearchPage.tsx
@@ -1,16 +1,11 @@
-import React, { useState } from 'react';
-import { Link } from 'react-router-dom';
-
-import Navbar from '../common/Navbar/Navbar';
+import React from 'react';
 
 import './SearchPage.css';
 
 const SearchPage = () => {
   return (
     <>
-      <div>
-        <Navbar />
-      </div>
+      <div></div>
     </>
   );
 };
