+++ client/src/components/common/Navbar/Navbar.css
@@ -1,83 +1,6 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  overflow-x: hidden;
-}
-
 .iou-logo-txt {
   color: #ffffff;
   font-size: 40px;
-}
-
-.main-content {
-  width: 100vw;
-  height: 100vh;
-  position: relative;
-  overflow-x: hidden;
-}
-.content {
-  width: 100%;
-  height: 100%;
-}
-
-.content-txt {
-  color: #ffffff;
-  font-size: 50px;
-  position: absolute;
-  top: 80px;
-  z-index: 50;
-  margin-top: 100px;
-  left: 50px;
-  text-align: left;
-}
-
-.register-btn {
-  color: #8f2a7e;
-  font-size: 25px;
-  font-weight: bold;
-  background-color: #ffffff;
-  height: 60px;
-  position: absolute;
-  top: 100px;
-  z-index: 50;
-  margin-top: 330px;
-  left: 50px;
-  border-radius: 10px;
-  border: none;
-  width: 300px;
-  cursor: pointer;
-}
-
-.public-favour-text-below {
-  margin-top: 20px;
-  color: #9b9b9b;
-}
-
-.register-btn:hover {
-  transition-duration: 0.3s;
-  background-color: #ff6122;
-  color: #ffffff;
-}
-.register-btn:focus {
-  outline: none;
-  color: #ffffff;
-  background-color: #ff6122;
-}
-
-.content-txt-below {
-  color: #ffffff;
-  position: absolute;
-  top: 50px;
-  z-index: 50;
-  font-size: 20px;
-  margin-top: 300px;
-  left: 50px;
-  text-align: left;
-}
-
-.public-title {
-  font-size: 50px;
 }
 
 .navbar-top {
@@ -273,100 +196,4 @@
   align-items: center;
   position: relative;
   justify-content: space-around;
-}
-
-.background-img {
-  position: relative;
-  z-index: 10;
-  left: 0;
-  top: 0;
-  overflow: hidden;
-}
-
-.isometric-img {
-  position: absolute;
-  z-index: 15;
-  right: 50px;
-  top: 100px;
-}
-
-.favours-card {
-  width: 1000px;
-  height: 100px;
-  margin: 50px auto;
-  border-radius: 10px;
-  border: 2px solid #dbdbdb;
-  -webkit-box-shadow: 0px 9px 20px 3px rgba(26, 26, 26, 0.16);
-  -moz-box-shadow: 0px 9px 20px 3px rgba(26, 26, 26, 0.16);
-  box-shadow: 0px 9px 20px 3px rgba(26, 26, 26, 0.16);
-}
-
-.favours-card-square-bg {
-  background-color: #ff6464;
-  height: 100%;
-  float: left;
-  width: 100px;
-  overflow: hidden;
-}
-
-.avatar {
-  width: 70px !important;
-  height: 100px !important;
-}
-
-.title-txt {
-  margin-top: 10px;
-  margin-left: 20px;
-  text-align: left;
-  float: left;
-  margin-right: 30px;
-  font-size: 15px;
-  color: #8b9ca8;
-}
-
-.card-text {
-  font-size: 17px;
-  color: #596773;
-  font-weight: bold;
-  display: inline;
-  float: left;
-  margin-left: 20px;
-}
-
-.points-text {
-  color: #15b987;
-  float: left;
-  text-align: left;
-  margin-left: 20px;
-  font-weight: bolder;
-}
-
-.do-favour-btn {
-  float: right;
-  display: inline;
-  transition-duration: 0.3s;
-  font-family: 'Montserrat', sans-serif;
-  background: #fe8e51;
-  border-radius: 5px;
-  width: 150px;
-  border: none;
-  margin-right: 40px;
-  height: 40px;
-  font-weight: 600;
-  color: white;
-  cursor: pointer;
-}
-
-.tag {
-  color: #4d6bff;
-  display: block;
-  border-radius: 30px;
-  font-weight: bold;
-  margin-top: 8px;
-  width: 100px;
-  height: 25px;
-  background-color: #afd2ff;
-}
-
-.tag-content {
 }
